pyproject.toml
[project]
name = "job-search-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.0.0",
    "docxtpl>=0.19.1",
    "folium>=0.19.4",
    "literalai>=0.0.623",
    "openai>=1.59.3",
    "pandas>=2.2.3",
    "pydeck>=0.9.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "supabase>=2.11.0",
]
